Change flag package to pflag. Add options to sort columns. Change constants to uint8 & update tests.

+++ calculate2_test.go
@@ -14,7 +14,7 @@ func TestCalculate2NameAsc(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		decimals uint
+		decimals uint8
 		expected string
 	}{
 		{decimals: 0, expected: `arch: amd64
@@ -79,7 +79,7 @@ func TestCalculate2NameDsc(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		decimals uint
+		decimals uint8
 		expected string
 	}{
 		{decimals: 0, expected: `arch: amd64
@@ -144,7 +144,7 @@ func TestCalculate2Maximum(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		decimals uint
+		decimals uint8
 		expected string
 	}{
 		{decimals: 0, expected: `arch: amd64
